// src/app/interface/calendario.interface.ts
// Generated by https://quicktype.io

export interface ResultCalendarios {
  ok:         boolean;
  msg:        string;
  calendario: Calendario[];
}
export interface ResultCalendarioID {
  ok:         boolean;
  msg:        string;
  calendario: Calendario;
}
export interface Calendario {
  id:          number;
  titulo:      string;
  descripcion: string;
  fecha:       string;
  ano:         string;
  mes:         string;
  img:         string;
}

export interface CalendarioForm{
  titulo:string,
  descripcion:string,
  fecha:string,
  imagen?:string,
}
export interface ImgCalendarioForm{
  imagen?:string
}